chore: First part of tiny Status refactoring
- Use file-scoped namespaces (needs project update)
- Rename StatusPartialTest to StatusTest and move to Rpc subdirectory
- Change namespace for StatusTest to Google.Rpc

No actual code changes in this commit.
