+++ Google.Api.CommonProtos/Google.Api.CommonProtos.csproj
@@ -5,6 +5,7 @@
   <PropertyGroup>
     <Version>2.12.0</Version>
     <TargetFrameworks>netstandard2.0;net461</TargetFrameworks>
+    <LangVersion>latest</LangVersion>
   </PropertyGroup>
 
   <!-- Packaging information -->
